refactor(search): extract join-inference tools to _JoinInferenceMixin
Moves the three join-discovery tools (get_join_candidates,
find_joinable_tables, find_joinable_across_profiles), their two private
helpers (_fetch_live_column_names, _compute_value_overlap_rows), and
the _VALID_JOIN_STRATEGIES constant out of ToolBox into a mixin module.

_sample_distinct_values graduates to _agent_tools_helpers so both the
mixin and the live-DB sampler can import it without a circular
dependency. Updates one test patch path that referenced the old
agent_tools location.
